tests/requirements.txt
ledger_app_clients.ethereum @ https://github.com/LedgerHQ/app-ethereum/archive/develop.zip#subdirectory=client
pytest
ragger[speculos,ledgercomm]
